Followers como iconos.
- Sistema para cargar formas/géneros de followers.
- Avance con sprites hasta Butterfree.

+++ include/pokemon_icon.h
@@ -3,10 +3,10 @@
 
 extern const struct SpritePalette gMonIconPaletteTable[];
 
-const u8 *GetMonIconTiles(u16 species, u32 personality);
+const u32 *GetMonIconTiles(u16 species, u32 personality);
 void TryLoadAllMonIconPalettesAtOffset(u16 offset);
 u8 GetValidMonIconPalIndex(u16 species);
-const u8 *GetMonIconPtr(u16 speciesId, u32 personality);
+const u32 *GetMonIconPtr(u16 speciesId, u32 personality);
 const u16 *GetValidMonIconPalettePtr(u16 speciesId);
 u16 GetIconSpecies(u16 species, u32 personality);
 u16 GetUnownLetterByPersonality(u32 personality);
